chore: update aws sdk to v3 and fix some rsc fails
