+++ apps/course-builder-web/package.json
@@ -32,7 +32,9 @@
     "@atlaskit/select": "^17.1.0",
     "@atlaskit/tokens": "^1.37.1",
     "@auth/core": "^0.28.0",
-    "@aws-sdk/client-s3": "^3.441.0",
+    "@aws-sdk/client-s3": "^3.525.0",
+    "@aws-sdk/client-textract": "^3.525.0",
+    "@aws-sdk/credential-providers": "^3.525.0",
     "@aws-sdk/s3-request-presigner": "^3.441.0",
     "@casl/ability": "^6.5.0",
     "@codemirror/lang-markdown": "^6.2.3",
@@ -94,7 +96,6 @@
     "@upstash/ratelimit": "^1.0.1",
     "@upstash/redis": "^1.28.4",
     "ai": "3.0.2",
-    "aws-sdk": "^2.1494.0",
     "axios": "^1.5.1",
     "base-64": "^1.0.0",
     "class-variance-authority": "^0.7.0",
