+++ apps/course-builder-web/src/app/tips/[slug]/page.tsx
@@ -5,7 +5,8 @@ import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { TipPlayer } from '@/app/tips/_components/tip-player'
 import { getAbility } from '@/lib/ability'
-import { getTip, type Tip } from '@/lib/tips'
+import { type Tip } from '@/lib/tips'
+import { getTip } from '@/lib/tips-query'
 import { getTranscript, getVideoResource } from '@/lib/video-resource'
 import { getServerAuthSession } from '@/server/auth'
 import { getOGImageUrlForResource } from '@/utils/get-og-image-url-for-resource'
